issue #30: introduced typeclasses for constructing, copying and transforming types;
