+++ src/main/scala/com/abraxas/slothql/Example.scala
@@ -3,110 +3,7 @@ package com.abraxas.slothql
 import shapeless._
 
 import com.abraxas.slothql.FragmentDefinitionHelper._
-import com.abraxas.slothql.util.Not
-
-
-trait OrElse[A, B] extends DepFn2[A, B]
-object OrElse {
-  type Aux[A, B, Out0] = OrElse[A, B] { type Out = Out0 }
-  def apply[A, B](implicit ev: A OrElse B): Aux[A, B, ev.Out] = ev
-
-  implicit def orElseA[A, B](implicit defined: A =:!= Nothing): Aux[A, B, A] = a.asInstanceOf[Aux[A, B, A]]
-  private lazy val a = new OrElse[Any, Any] {
-    type Out = Any
-    def apply(a: Any, b: Any): Any = a
-  }
-
-  implicit def orElseB[A, B](implicit undefined: A =:= Nothing): Aux[A, B, B] = b.asInstanceOf[Aux[A, B, B]]
-  private lazy val b = new OrElse[Any, Any] {
-    type Out = Any
-    def apply(a: Any, b: Any): Any = b
-  }
-}
-
-
-trait MapOrElse[HF <: Poly1, A, B] extends DepFn2[A, B]
-object MapOrElse {
-  type Aux[HF <: Poly1, A, B, Out0] = MapOrElse[HF, A, B] { type Out = Out0 }
-  def apply[HF <: Poly1, A, B](implicit moe: MapOrElse[HF, A, B]): Aux[HF, A, B, moe.Out] = moe
-
-  implicit def mapOrElseA[HF <: Poly1, A, B, A0](
-    implicit
-    defined: A =:!= Nothing,
-    hf: poly.Case1.Aux[HF, A0, A]
-  ): Aux[HF, A0, B, A] =
-    new MapOrElse[HF, A0, B] {
-      type Out = A
-      def apply(a0: A0, b: B): A = hf(a0)
-    }
-
-  implicit def mapOrElseB[HF <: Poly1, A, B](implicit undefined: A =:= Nothing): Aux[HF, A, B, B] = b.asInstanceOf[Aux[HF, A, B, B]]
-  private lazy val b = new MapOrElse[Poly1, Any, Any] {
-    type Out = Any
-    def apply(a0: Any, b: Any): Any = b
-  }
-
-}
-
-
-trait Construct[A, Args <: HList] extends DepFn1[Args] { type Out <: A }
-object Construct {
-  type Aux[A, Args <: HList, Out0 <: A] = Construct[A, Args] { type Out = Out0 }
-
-  def apply[A]: Builder[A] = Builder.asInstanceOf[Builder[A]]
-
-  protected class Builder[A] extends ProductArgs {
-    def applyProduct[Args <: HList](args: Args)(implicit build: Construct[A, Args]): Aux[A, Args, build.Out] = build
-  }
-  private object Builder extends Builder[Any]
-}
-
-trait Copy[A, Changes <: HList] extends DepFn2[A, Changes]
-object Copy {
-  type Aux[A, Changes <: HList, R] = Copy[A, Changes] { type Out = R }
-
-  def apply[A](a: A): Builder[A] = new Builder[A](a)
-
-  protected class Builder[A](a: A) extends RecordArgs {
-    def applyRecord[Changes <: HList](changes: Changes)(implicit copy: Copy[A, Changes]): Aux[A, Changes, copy.Out] = copy
-  }
-
-  implicit def copyIdentity[A]: Copy.Aux[A, HNil, A] = _copyIdentity.asInstanceOf[Copy.Aux[A, HNil, A]]
-  private lazy val _copyIdentity = new Copy[Any, HNil] {
-    type Out = Any
-    def apply(t: Any, u: HNil): Any = t
-  }
-}
-
-trait Transform[A, HF <: Poly1] extends DepFn1[A]
-object Transform {
-  type Aux[A, HF <: Poly1, R] = Transform[A, HF] { type Out = R }
-
-  def apply[A, HF <: Poly1](a: A, hf: HF)(implicit transform: Transform[A, HF]): Aux[A, HF, transform.Out] = transform
-
-  @deprecated("recursion is deactivated")
-  implicit def transformThis[A, HF <: Poly1, R0](implicit hf: poly.Case1.Aux[HF, A, R0]): Aux[A, HF, R0] =
-    new Transform[A, HF] {
-      type Out = R0
-      def apply(t: A): R0 = hf(t)
-    }
-
-/* TODO: recurse
-  implicit def transformThis[A, HF <: Poly1, R0](
-    implicit
-    hf: poly.Case1.Aux[HF, A, R0],
-    next: Transform[R0, HF]
-  ): Aux[A, HF, next.Out] =
-    new Transform[A, HF] {
-      type Out = next.Out
-      def apply(t: A): next.Out = next(hf(t))
-    }
-*/
-
-}
-
-
-// // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // //
+import com.abraxas.slothql.util._
 
 
 
@@ -128,14 +25,18 @@ object Lit{
     }
 }
 
-//  /** ¡Template! */
-//  trait Call0[A] extends Expr[A] {
-//    type Func = fromNarrow[String]
-//    type Params <: HList
-//
-//    val func: Func As String
-//    val params: FromHList[Params, LiftKnown.type, List[Known[Expr[_]]]]
-//  }
+
+// // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // // //
+
+
+/** ¡Template! */
+trait Call0[A] extends Expr[A] {
+  type Func <: String
+  type Params <: HList
+
+  val func: String fromNarrow Func
+  val params: FromHList[Params, LiftKnown.type, List[Known[Expr[_]]]]
+}
 
 object Common {
   def undefined: Nothing = sys.error("undefined")
